+++ src/main/java/bowling/domain/frame/Frame.java
@@ -1,4 +1,6 @@
-package bowling.domain;
+package bowling.domain.frame;
+
+import bowling.domain.state.State;
 
 public interface Frame {
     public static final String WRONG_FALLED_PIN = "넘어뜨린 핀의 개수가 알맞지 않습니다.";
