+++ src/main/java/bowling/domain/state/Playing.java
@@ -1,4 +1,4 @@
-package bowling.domain;
+package bowling.domain.state;
 
 public class Playing implements State{
