+++ src/main/java/bowling/domain/state/State.java
@@ -1,4 +1,4 @@
-package bowling.domain;
+package bowling.domain.state;
 
 public interface State {
     public State play(int falledPin);
